Split widget: batch row
