+++ extension/devtools/app/lib/main.dart
@@ -57,6 +57,7 @@ part 'widgets/computed_row.dart';
 part 'widgets/computed_detail.dart';
 part 'widgets/batch_list.dart';
 part 'widgets/batch_header_row.dart';
+part 'widgets/batch_row.dart';
 part 'widgets/panel_placeholder.dart';
 part 'widgets/panel_scroll_view.dart';
 part 'widgets/panel_state_cards.dart';
@@ -984,102 +985,6 @@ class _BatchingPanel extends StatelessWidget {
             );
           },
         ),
-      ),
-    );
-  }
-}
-
-class _BatchRow extends StatelessWidget {
-  const _BatchRow({required this.batch, required this.isCompact});
-
-  final BatchSample batch;
-  final bool isCompact;
-
-  @override
-  Widget build(BuildContext context) {
-    final textTheme = Theme.of(context).textTheme;
-    return _GlassCard(
-      padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 12),
-      child: DefaultTextStyle.merge(
-        style: textTheme.bodyMedium,
-        child: isCompact
-            ? Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Text('Batch #${batch.id}'),
-                  const SizedBox(height: 6),
-                  Wrap(
-                    spacing: 8,
-                    runSpacing: 8,
-                    children: [
-                      _GlassPill(
-                        padding: const EdgeInsets.symmetric(
-                          horizontal: 10,
-                          vertical: 6,
-                        ),
-                        child: Text('Depth ${batch.depth}'),
-                      ),
-                      _GlassPill(
-                        padding: const EdgeInsets.symmetric(
-                          horizontal: 10,
-                          vertical: 6,
-                        ),
-                        child: Text('${batch.writeCount} writes'),
-                      ),
-                      _GlassPill(
-                        padding: const EdgeInsets.symmetric(
-                          horizontal: 10,
-                          vertical: 6,
-                        ),
-                        child: Text('${batch.durationMs}ms'),
-                      ),
-                      _GlassPill(
-                        padding: const EdgeInsets.symmetric(
-                          horizontal: 10,
-                          vertical: 6,
-                        ),
-                        child: Text(_formatAge(batch.endedAt)),
-                      ),
-                    ],
-                  ),
-                ],
-              )
-            : Row(
-                children: [
-                  Expanded(
-                    flex: 2,
-                    child: Text(
-                      'Batch #${batch.id}',
-                      textAlign: TextAlign.center,
-                    ),
-                  ),
-                  Expanded(
-                    flex: 2,
-                    child: Text('${batch.depth}', textAlign: TextAlign.center),
-                  ),
-                  Expanded(
-                    flex: 2,
-                    child: Text(
-                      '${batch.writeCount}',
-                      textAlign: TextAlign.center,
-                    ),
-                  ),
-                  Expanded(
-                    flex: 2,
-                    child: Text(
-                      '${batch.durationMs}ms',
-                      textAlign: TextAlign.center,
-                    ),
-                  ),
-                  Expanded(
-                    flex: 3,
-                    child: Text(
-                      _formatAge(batch.endedAt),
-                      textAlign: TextAlign.center,
-                    ),
-                  ),
-                ],
-              ),
       ),
     );
   }
